requirements.txt
# DistriStore Dependencies
# Phase 1: Core
PyYAML>=6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Phase 2: Cryptography
pycryptodome>=3.19.0

# Phase 5: SDK
httpx>=0.25.0

# Phase 6: Frontend build tool (optional, frontend uses npm)
